Updated failing execution tests: syntaxTests/{viewPureChecker, parsing}.

+++ test/libsolidity/syntaxTests/parsing/exp_expression.sol
@@ -1,8 +1,8 @@
 contract test {
-    function fun(uint256 a) public {
-        uint256 x = 3 ** a;
+    function fun(uint a) public {
+        uint x = 3 ** a;
     }
 }
 // ----
-// Warning 2072: (61-70): Unused local variable.
-// Warning 2018: (20-86): Function state mutability can be restricted to pure
+// Warning 2072: (58-64): Unused local variable.
+// Warning 2018: (20-80): Function state mutability can be restricted to pure
